requirements.txt
google-generativeai==0.3.2
python-dotenv==1.0.0
pillow==10.0.0
piexif==1.1.3
